add image

+++ server/src/controllers/ProductController.ts
@@ -2,11 +2,19 @@ import { Request, Response } from 'express';
 import ProductService from '../services/ProductService';
 import { successResponse, errorResponse } from '../utils/response';
 import { AuthRequest } from '../middleware/AuthMiddleWare';
+import cloudinary from '../utils/cloudinary';
 
-
-export const createProductController = (req: AuthRequest, res: Response) => {
+export const createProductController = async(req: Request, res: Response) => {
     try {
-        const product = ProductService.createProductService(req.body);
+        let imageUrl=""
+        let imageId=""  
+
+        if(req.file){
+            const res= await cloudinary.uploader.upload(req.file.path)
+            imageUrl=res.secure_url
+            imageId=res.public_id
+        }
+        const product = ProductService.createProductService({...req.body,imageUrl,imageId});
         return successResponse(res, product, "Product created successfully", 201);
     } catch (error) {
         return errorResponse(res, error, "Failed to create product", 400);
@@ -69,19 +77,48 @@ export const getBestSellersProductsController = (req: Request, res: Response) =>
 
 }
 
-export const updateProductController = (req: AuthRequest, res: Response) => {
+export const updateProductController = async(req: Request, res: Response) => {
     try {
         const id = Number(req.params.id);
-        const product = ProductService.updateProductService(id, req.body);
+        const existing =ProductService.getProductByIdService(id)
+        if(!existing){
+            return errorResponse(res,null,"Product not found",404)
+        }
+        let imageUrl=existing.imageUrl
+        let imageId=existing.imageId
+
+        if(req.file){
+            //del.old
+            if(existing.imageId){
+                await cloudinary.uploader.destroy(existing.imageId)
+            }
+            //up new
+            const res =await cloudinary.uploader.upload(req.file.path)
+
+            imageUrl=res.secure_url
+            imageId=res.public_id
+        }
+
+        const product = ProductService.updateProductService(id,{... req.body,imageUrl,imageId});
         return successResponse(res, product, "Product updated successfully");
     } catch (error) {
         return errorResponse(res, error, "Failed to update product", 400);
     }
 };
 
-export const deleteProductController = (req: AuthRequest, res: Response) => {
+export const deleteProductController =async (req: Request, res: Response) => {
     try {
         const id = Number(req.params.id);
+        const existing =ProductService.getProductByIdService(id)
+        if(!existing){
+            return errorResponse(res,null,"Product not found",404)
+        }
+        //del(cloud)
+        if(existing.imageId){
+            await cloudinary.uploader.destroy(existing.imageId)
+
+        }
+
         const result = ProductService.deleteProductService(id);
         return successResponse(res, result, "Product deleted successfully");
     } catch (error) {
